Create projects component
