+++ src/App.js
@@ -9,7 +9,7 @@ import Menu from './components/Menu/Menu';
 import Welcome from './components/Welcome/Welcome';
 import About from './components/About/About';
 import Skills from './components/Skills/Skills';
-import Projects from './components/Projects';
+import Projects from './components/Projects/Projects';
 
 export default function App() {
 	const [hasLoaded, setHasLoaded] = useState(false);
