+++ src/sass/main.scss
@@ -5,6 +5,6 @@
 @import '../components/Welcome/welcome';
 @import '../components/About/about';
 @import '../components/Skills/skills';
-@import 'components/projects';
+@import '../components/Projects/projects';
 @import 'base/global';
 @import '~bootstrap/scss/bootstrap';
